Refactor database schema to use UUIDs for user and cart IDs, update SQL queries accordingly, and improve consistency in API parameter types

+++ server/internal/db/query/category_product.sql
@@ -33,7 +33,7 @@ WHERE
 
 -- name: GetMaxSortOrderInCollection :one
 SELECT
-    max(sort_order)::smallint
+    max(sort_order)
 FROM
     category_products
 WHERE
